Add LICENSE and copyright notices
This commit adds an MIT license to the project, same to the original
ajmyyra/ambassador-auth-oidc project.
A copyright notice for Arrikto has been added on the license.

All code that was reused from the previous project is placed inside
util.go.
